pyproject.toml
[project]
name = "missionagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
